refactor: change PartyID to uint32 type alias
Replace "type PartyID uint32" (named type) with "type PartyID = uint32"
(type alias). Since PartyID is now identical to uint32, remove the
now-unnecessary conversion helpers and casts:

- Remove transcript.Uint32s() generic helper and uint32Value constraint
- Remove tss.Uint32sToPartySet() conversion function
- Remove ~63 unnecessary uint32(partyID) casts throughout the codebase
- Remove unnecessary tss.PartyID() constructor casts
- Drop TestUint32sConvertsNamedValues and uint32Bytes test helper

Update stale doc comments in internal/wire that referenced PartyID
as a named type example.

All tests pass. No wire format or transcript hash changes.

+++ broadcast.go
@@ -31,7 +31,7 @@ func AckDigest(protocol ProtocolID, sessionID SessionID, round uint8, from Party
 	t.AppendString("protocol", string(protocol))
 	t.AppendBytes("session_id", sessionID[:])
 	t.AppendUint8("round", round)
-	t.AppendUint32("from", uint32(from))
+	t.AppendUint32("from", from)
 	t.AppendString("payload_type", string(payloadType))
 	t.AppendBytes("payload_hash", payloadHash[:])
 	t.AppendBytes("transcript_hash", transcriptHash[:])
